requirements.txt
requests
pandas
kafka-python